Refactor Collection/Lst/LstBase: No more virtual inheritance
